Adding Chart Logic + Documentation
* Adding DockerFile
* Introducing the Chart system
* Adding Chart Controller + Services
* Adding Tests to validar Chart System
* Adding Documentation and Diagrams

During the Documentation implementation i realized that a Chart System might be required in order to add
products or remove those from the chart/basket so i decided to add a DynamicSupervisor that will Spawn workers (Individual charts)
based in the config.

In this way we can have multiples charts/users operating simultaneously and they don't need to make a queue to wait for someone finish their purchase action.